delete value.go
